WIP multiprocessing logging formatting
