+++ pandas_to_postgres/__init__.py
@@ -1,13 +1,7 @@
 from .copy_df import DataFrameCopy
-from .copy_hdf import HDFTableCopy, SmallHDFTableCopy, BigHDFTableCopy
+from .copy_hdf import HDFTableCopy, SmallHDFTableCopy, BigHDFTableCopy, HDFMetadata
 from .hdf_to_postgres import (
     hdf_to_postgres,
     create_hdf_table_objects,
 )
-from .utilities import (
-    logger,
-    HDFMetadata,
-    create_file_object,
-    df_generator,
-    cast_pandas,
-)
+from .utilities import create_file_object, df_generator, cast_pandas
