+++ pandas_to_postgres/copy_hdf.py
@@ -1,6 +1,8 @@
-from .utilities import create_file_object, df_generator, logger, cast_pandas
-from ._base_copy import BaseCopy
+import logging
 import pandas as pd
+from .utilities import create_file_object, df_generator, cast_pandas
+from ._base_copy import BaseCopy
+from collections import defaultdict
 
 
 class HDFTableCopy(BaseCopy):
@@ -86,33 +88,35 @@ class HDFTableCopy(BaseCopy):
         data_formatter_kwargs: list of kwargs to pass to data_formatters functions
         """
         if self.hdf_tables is None:
-            logger.warn("No HDF table found for SQL table {}".format(self.sql_table))
+            self.logger.warn(
+                "No HDF table found for SQL table {}".format(self.sql_table)
+            )
             return
 
         for hdf_table in self.hdf_tables:
-            logger.info("*** {} ***".format(hdf_table))
+            self.logger.info("*** {} ***".format(hdf_table))
 
-            logger.info("Reading HDF table")
+            self.logger.info("Reading HDF table")
             df = pd.read_hdf(self.file_name, key=hdf_table)
             self.rows += len(df)
 
             data_formatter_kwargs["hdf_table"] = hdf_table
-            logger.info("Formatting data")
+            self.logger.info("Formatting data")
             df = self.data_formatting(
                 df, functions=data_formatters, **data_formatter_kwargs
             )
 
-            logger.info("Creating generator for chunking dataframe")
-            for chunk in df_generator(df, self.csv_chunksize):
+            self.logger.info("Creating generator for chunking dataframe")
+            for chunk in df_generator(df, self.csv_chunksize, logger=self.logger):
 
-                logger.info("Creating CSV in memory")
+                self.logger.info("Creating CSV in memory")
                 fo = create_file_object(chunk)
 
-                logger.info("Copying chunk to database")
+                self.logger.info("Copying chunk to database")
                 self.copy_from_file(fo)
                 del fo
             del df
-        logger.info("All chunks copied ({} rows)".format(self.rows))
+        self.logger.info("All chunks copied ({} rows)".format(self.rows))
 
 
 class SmallHDFTableCopy(HDFTableCopy):
@@ -172,29 +176,29 @@ class SmallHDFTableCopy(HDFTableCopy):
         data_formatter_kwargs: list of kwargs to pass to data_formatters functions
         """
         if self.hdf_tables is None:
-            logger.warn("No HDF table found for SQL table {self.sql_table}")
+            self.logger.warn("No HDF table found for SQL table {self.sql_table}")
             return
 
         for hdf_table in self.hdf_tables:
-            logger.info("*** {} ***".format(hdf_table))
-            logger.info("Reading HDF table")
+            self.logger.info("*** {} ***".format(hdf_table))
+            self.logger.info("Reading HDF table")
             df = pd.read_hdf(self.file_name, key=hdf_table)
             self.rows += len(df)
 
             data_formatter_kwargs["hdf_table"] = hdf_table
-            logger.info("Formatting data")
+            self.logger.info("Formatting data")
             df = self.data_formatting(
                 df, functions=data_formatters, **data_formatter_kwargs
             )
 
-            logger.info("Creating CSV in memory")
+            self.logger.info("Creating CSV in memory")
             fo = create_file_object(df)
 
-            logger.info("Copying table to database")
+            self.logger.info("Copying table to database")
             self.copy_from_file(fo)
             del df
             del fo
-        logger.info("All chunks copied ({} rows)".format(self.rows))
+        self.logger.info("All chunks copied ({} rows)".format(self.rows))
 
 
 class BigHDFTableCopy(HDFTableCopy):
@@ -257,11 +261,13 @@ class BigHDFTableCopy(HDFTableCopy):
         data_formatter_kwargs: list of kwargs to pass to data_formatters functions
         """
         if self.hdf_tables is None:
-            logger.warn("No HDF table found for SQL table {}".format(self.sql_table))
+            self.logger.warn(
+                "No HDF table found for SQL table {}".format(self.sql_table)
+            )
             return
 
         for hdf_table in self.hdf_tables:
-            logger.info("*** {} ***".format(hdf_table))
+            self.logger.info("*** {} ***".format(hdf_table))
 
             with pd.HDFStore(self.file_name) as store:
                 nrows = store.get_storer(hdf_table).nrows
@@ -275,26 +281,85 @@ class BigHDFTableCopy(HDFTableCopy):
             start = 0
 
             for i in range(n_chunks):
-                logger.info("*** HDF chunk {i + 1} of {} ***".format(n_chunks))
-                logger.info("Reading HDF table")
+                self.logger.info("*** HDF chunk {i + 1} of {} ***".format(n_chunks))
+                self.logger.info("Reading HDF table")
                 stop = min(start + self.hdf_chunksize, nrows)
                 df = pd.read_hdf(self.file_name, key=hdf_table, start=start, stop=stop)
 
                 start += self.hdf_chunksize
 
                 data_formatter_kwargs["hdf_table"] = hdf_table
-                logger.info("Formatting data")
+                self.logger.info("Formatting data")
                 df = self.data_formatting(
                     df, functions=data_formatters, **data_formatter_kwargs
                 )
 
-                logger.info("Creating generator for chunking dataframe")
-                for chunk in df_generator(df, self.csv_chunksize):
-                    logger.info("Creating CSV in memory")
+                self.logger.info("Creating generator for chunking dataframe")
+                for chunk in df_generator(df, self.csv_chunksize, logger=self.logger):
+                    self.logger.info("Creating CSV in memory")
                     fo = create_file_object(chunk)
 
-                    logger.info("Copying chunk to database")
+                    self.logger.info("Copying chunk to database")
                     self.copy_from_file(fo)
                     del fo
                 del df
-        logger.info("All chunks copied ({} rows)".format(self.rows))
+        self.logger.info("All chunks copied ({} rows)".format(self.rows))
+
+
+class HDFMetadata(object):
+    """Collect applicable metadata from HDFStore to use when running copy"""
+
+    def __init__(
+        self,
+        file_name,
+        keys=None,
+        chunksize=10 ** 7,
+        metadata_attr=None,
+        metadata_keys=[],
+    ):
+        self.file_name = file_name
+        self.chunksize = chunksize
+        self.sql_to_hdf = defaultdict(set)
+        self.metadata_vars = defaultdict(dict)
+        self.logger = logging.getLogger("HDFMetadata")
+
+        """
+        Parameters
+        ----------
+        file_name: str
+            path to hdf file to copy from
+        keys: list of strings
+            HDF keys to copy data from
+        chunksize: int
+            Maximum rows read from an hdf file into a pandas dataframe when using
+            the BigTable protocol
+        metadata_attr: str
+            :ocation of relevant metadata in store.get_storer().attrs
+        metadata_keys: list of strings
+            Keys to get from metadata store
+        """
+
+        with pd.HDFStore(self.file_name, mode="r") as store:
+            self.keys = keys or store.keys()
+
+            if metadata_attr:
+                for key in self.keys:
+                    try:
+                        metadata = store.get_storer(key).attrs[metadata_attr]
+                        self.logger.info("Metadata: {}".format(metadata))
+                    except (AttributeError, KeyError):
+                        if "/meta" not in key:
+                            self.logger.info(
+                                "No metadata found for key '{}'. Skipping".format(key)
+                            )
+                        continue
+
+                    for mkey in metadata_keys:
+                        self.metadata_vars[mkey][key] = metadata.get(mkey)
+
+                    sql_table = metadata.get("sql_table_name")
+
+                    if sql_table:
+                        self.sql_to_hdf[sql_table].add(key)
+                    else:
+                        self.logger.warn("No SQL table name found for {}".format(key))
